Add partitioning calls
