+++ k8s/production/stock.yml
@@ -34,6 +34,8 @@ spec:
               value: INFO
             - name: SPRING_KAFKA_BOOTSTRAP_SERVERS
               value: wdm-kafka-kafka:9092
+            - name: NUM_PARTITIONS
+              value: 2
             - name: POD_NAME
               valueFrom:
                 fieldRef:
